refactor: internal refactors
